requirements.txt
aniso8601==7.0.0
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.4
parso==0.7.0
promise==2.3
psycopg2==2.8.5
pyparsing==2.4.7
python-mpd2==1.0.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
SQLAlchemy==1.3.17
Werkzeug==1.0.1
